requirements.txt
openai==0.27.8
chromadb==0.4.3
langchain==0.0.247
sentence-transformers==2.2.2
# https://github.com/dagger/dagger/issues/5593
#dagger-io==0.8.1
dagger-io==0.6.4
